Refactor to remove more package tangle.
