+++ project-set/components/client-auth/src/main/java/com/rackspace/papi/components/clientauth/ClientAuthenticationFilter.java
@@ -3,7 +3,7 @@ package com.rackspace.papi.components.clientauth;
 import com.rackspace.papi.components.clientauth.config.ClientAuthConfig;
 import com.rackspace.papi.filter.logic.impl.FilterLogicHandlerDelegate;
 import com.rackspace.papi.service.config.ConfigurationService;
-import com.rackspace.papi.service.context.ContextAdapter;
+import com.rackspace.papi.service.context.common.ContextAdapter;
 import com.rackspace.papi.service.context.ServletContextHelper;
 import com.rackspace.papi.service.datastore.Datastore;
 import com.rackspace.papi.service.datastore.DatastoreService;
