+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/filter/PowerFilter.java
@@ -10,7 +10,7 @@ import com.rackspace.papi.domain.ServicePorts;
 import com.rackspace.papi.model.Node;
 import com.rackspace.papi.model.ReposeCluster;
 import com.rackspace.papi.model.SystemModel;
-import com.rackspace.papi.service.context.ContextAdapter;
+import com.rackspace.papi.service.context.common.ContextAdapter;
 import com.rackspace.papi.service.context.ServletContextHelper;
 import com.rackspace.papi.service.deploy.ApplicationDeploymentEvent;
 import com.rackspace.papi.service.event.PowerFilterEvent;
@@ -66,7 +66,7 @@ public class PowerFilter extends ApplicationContextAwareFilter {
                     newFilterChain);
          }
       }
-   };
+   }
 
    private class SystemModelConfigListener implements UpdateListener<SystemModel> {
