+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/context/common/ContextAdapter.java
@@ -1,7 +1,8 @@
-package com.rackspace.papi.service.context;
+package com.rackspace.papi.service.context.common;
 
 import com.rackspace.papi.service.classloader.ClassLoaderManagerService;
 import com.rackspace.papi.service.config.ConfigurationService;
+import com.rackspace.papi.service.context.ServiceContext;
 import com.rackspace.papi.service.context.container.ContainerConfigurationService;
 import com.rackspace.papi.service.datastore.DatastoreService;
 import com.rackspace.papi.service.event.common.EventService;
