+++ project-set/core/core-lib/src/main/java/com/rackspace/papi/service/context/spring/SpringContextAdapter.java
@@ -2,7 +2,7 @@ package com.rackspace.papi.service.context.spring;
 
 import com.rackspace.papi.service.classloader.ClassLoaderManagerService;
 import com.rackspace.papi.service.config.ConfigurationService;
-import com.rackspace.papi.service.context.ContextAdapter;
+import com.rackspace.papi.service.context.common.ContextAdapter;
 import com.rackspace.papi.service.context.ServiceContext;
 import com.rackspace.papi.service.context.container.ContainerConfigurationService;
 import com.rackspace.papi.service.context.impl.*;
